:hammer: refactor : 링크 복사 컴포넌트 분리
